refactor: simplify and consolidate some test code
